# openapi.yaml
openapi: 3.0.3
info:
  title: 焚信 BurnMsg API
  version: 2.0.0
  description: |-
    焚信（BurnMsg）是一款企业内部加密即时通讯（IM）应用，支持云聊天与秘密聊天、阅后即焚消息。本文件为其后端（NestJS + TypeORM + MySQL）的标准 OpenAPI 3.0.3 接口文档。

    【账号体系】
    - 不开放自助注册，所有账号由管理后台统一开通（POST /accounts，仅 admin）。
    - 用户名即公司内部手机号（1 开头 11 位数字）。
    - 新开通 / 被重置密码的账号 force_change_pwd=true，前端应强制跳转修改密码，修改成功后该标记置为 false。

    【认证方式】
    - 受保护接口在请求头携带：Authorization: Bearer <access_token>。
    - 双 token 机制：access_token（JWT_SECRET 签名，默认有效期 7200 秒）+ refresh_token（独立的 JWT_REFRESH_SECRET 签名，默认有效期 2592000 秒 / 30 天，payload 带 type=refresh 标记防混用）。
    - access_token 过期后用 refresh_token 调 POST /auth/refresh-token 换取新 token 对。
    - 即时吊销：JWT 守卫除验签外，每个受保护请求都会实时查库校验「账号仍为 active」且「签发 token 的设备记录仍存在」；账号被停用或设备被下线后，已签发的 token 立即失效（全端即时强制下线）。

    【统一响应格式】
    - 成功：{ "code": 0, "message": "ok 或各接口自定义提示语", "data": <业务数据，无内容时为 null> }
    - 失败：{ "code": <HTTP 状态码>, "message": "<错误描述；参数校验失败时为多条错误以 '; ' 拼接>" }

    【通用约定】
    - 分页参数：page（默认 1）、pageSize（默认 20，最大 200；通讯录列表默认 50）。
    - 请求中的时间字段使用 ISO8601 字符串；响应中的时间为服务端 datetime。
    - 所有业务 ID 均为 UUID v4 字符串。
    - 角色：user / admin。标注「仅管理员」的接口仅 admin 可调用，普通用户返回 403。
    - 强制改密拦截：force_change_pwd=true 的账号（新开通 / 被重置密码后），除 POST /auth/change-password 与 GET /auth/profile 外，调用其他任何受保护接口一律返回 403（FORCE_CHANGE_PASSWORD），前端应引导其先完成改密再进入主界面。
    - 软删除账号：被 DELETE /accounts/{id} 软删的账号立即吊销全部 token 并强制下线，其历史消息保留、前端统一展示「已注销用户」；账号列表默认不含已软删账号，加 show_deleted=1 可查看。
servers:
  - url: http://localhost:9091/api/v1
...
security:
  - bearerAuth: []
...
paths:
...
  /accounts:
    post:
      tags: [账号]
      summary: 开通单个账号
      description: |-
        【用途】管理员为员工开通单个账号（本系统不开自助注册的唯一开户入口之一）。

        【调用时机】管理后台「添加成员」操作时。

        【业务规则与注意事项】
        - 仅 admin 角色可调用，普通用户返回 403。
        - 手机号必须唯一，重复开通返回 409。
        - password 留空时服务端自动生成 12 位随机密码（去除易混淆字符），并在响应 initial_password 中返回一次，请立即转交员工并妥善保管；指定了 password 时 initial_password 返回的即为所指定值。
        - 新账号初始：role=user、status=active、force_change_pwd=true（员工首登强制改密）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, display_name]
              properties:
                phone:
                  type: string
                  pattern: '^1\d{10}$'
                  description: 手机号（即用户名），1 开头共 11 位数字，全局唯一
                  example: '13900139001'
                display_name:
                  type: string
                  maxLength: 100
                  description: 姓名/显示名，最长 100 字符
                  example: 张三
                department:
                  type: string
                  maxLength: 100
                  description: 部门（可选），最长 100 字符
                  example: 工程部
                password:
                  type: string
                  description: 初始密码（可选）；留空则自动生成 12 位随机密码并在响应中返回一次
                  example: 'Init@2026'
      responses:
        '200':
          description: 开通成功（message 为「账号开通成功」）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 账号开通成功 }
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/SafeUser'
                      initial_password:
                        type: string
                        description: 初始密码（未指定 password 时为服务端生成的随机密码，仅此一次返回，请妥善转交）
                        example: 'aB3#eF7kM9pQ'
        '400':
          description: 参数校验失败（手机号格式不正确、display_name 缺失或超长等）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非 admin 角色（Admin access required）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: 该手机号已开通账号
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      tags: [账号]
      summary: 账号分页列表
      description: |-
        【用途】管理员分页查询全部账号（含已停用），支持按关键字、部门、状态过滤。

        【调用时机】管理后台「成员管理」列表页加载与筛选时。

        【业务规则与注意事项】
        - 仅 admin 角色可调用，普通用户返回 403。
        - keyword 为手机号精确匹配 OR 姓名模糊匹配（LIKE %keyword%）。
        - 按创建时间倒序排列。
        - 响应分页结构为 { data: SafeUser[], total }（注意字段名为 data，不是 list）。
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
          description: 页码，默认 1
          example: 1
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
          description: 每页条数，默认 20，最大 200
          example: 20
        - name: keyword
          in: query
          schema: { type: string }
          description: 关键字，手机号精确匹配或姓名模糊匹配
          example: 张
        - name: department
          in: query
          schema: { type: string }
          description: 按部门精确过滤
          example: 工程部
        - name: status
          in: query
          schema: { type: string, enum: [active, disabled] }
          description: 按状态过滤：active 正常 / disabled 已停用
        - name: show_deleted
          in: query
          schema: { type: string, enum: ['1'] }
          description: 传 1 时在结果中包含已软删除账号（deleted_at 非空，前端应置灰展示「已注销」）；默认不含
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: ok }
                  data:
                    type: object
                    properties:
                      data:
                        type: array
                        description: 当前页账号列表（SafeUser）
                        items:
                          $ref: '#/components/schemas/SafeUser'
                      total:
                        type: integer
                        description: 满足条件的账号总数
                        example: 137
        '400':
          description: 分页参数校验失败（page/pageSize 非整数或越界）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非 admin 角色（Admin access required）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /accounts/batch:
    post:
      tags: [账号]
      summary: JSON 批量开通账号（1~500 条）
      description: |-
        【用途】管理员以 JSON 数组方式批量开通账号。

        【调用时机】管理后台粘贴/录入批量名单后提交时；Excel 文件导入请用 POST /accounts/import。

        【业务规则与注意事项】
        - 仅 admin 角色可调用，普通用户返回 403。
        - 数组长度 1~500，逐条独立开通：单条失败（如手机号已存在）不影响其它条目，失败项在结果中带 error 原因。
        - 每条账号均自动生成 12 位随机初始密码（批量接口不支持指定密码），成功项的 password 字段返回该初始密码，请妥善导出转交。
        - 新账号初始 force_change_pwd=true（首登强制改密）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accounts]
              properties:
                accounts:
                  type: array
                  minItems: 1
                  maxItems: 500
                  description: 待开通账号数组，1~500 条
                  items:
                    type: object
                    required: [phone, display_name]
                    properties:
                      phone:
                        type: string
                        pattern: '^1\d{10}$'
                        description: 手机号，1 开头共 11 位数字
                        example: '13900139002'
                      display_name:
                        type: string
                        maxLength: 100
                        description: 姓名，最长 100 字符
                        example: 李四
                      department:
                        type: string
                        maxLength: 100
                        description: 部门（可选）
                        example: 市场部
      responses:
        '200':
          description: 批量导入完成（message 为「批量导入完成」），data 为逐条结果明细
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 批量导入完成 }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BatchResult'
        '400':
          description: 参数校验失败（数组为空/超过 500 条、某条手机号格式不正确等）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非 admin 角色（Admin access required）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /accounts/import:
    post:
      tags: [账号]
      summary: Excel 批量导入账号（xlsx，最大 5MB）
      description: |-
        【用途】管理员上传 Excel（xlsx）文件批量导入账号。

        【调用时机】管理后台选择既有 Excel 名单文件导入时；JSON 方式请用 POST /accounts/batch。

        【业务规则与注意事项】
        - 仅 admin 角色可调用，普通用户返回 403。
        - multipart/form-data 上传，文件字段名 file，仅解析第一个工作表，文件最大 5MB（超出返回 413）。
        - 列名支持中英文：phone/手机号、display_name/姓名、department/部门；phone 与 display_name 必填（无效行被自动过滤），department 可选。
        - 过滤后无有效账号行返回 409；Excel 文件解析失败返回 409。
        - 逐条独立开通，单条失败不影响其它条目；成功账号自动生成随机初始密码并在结果 password 字段返回，新账号 force_change_pwd=true。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: xlsx 文件（最大 5MB）；列名支持 phone/手机号、display_name/姓名、department/部门
      responses:
        '200':
          description: 导入完成（message 为「Excel 导入完成」），data 为逐条结果明细
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: Excel 导入完成 }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BatchResult'
        '400':
          description: 未接收到文件（未选择文件或字段名不是 file）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非 admin 角色（Admin access required）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Excel 文件解析失败，请检查文件格式；或 Excel 中无有效账号行
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '413':
          description: 文件超过 5MB 限制
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /accounts/departments:
    get:
      tags: [账号]
      summary: 部门去重列表
      description: |-
        【用途】获取全公司账号中出现过的部门名（去重、升序），用于筛选下拉框。

        【调用时机】管理后台成员列表页的部门筛选器初始化时。

        【业务规则与注意事项】
        - 仅 admin 角色可调用，普通用户返回 403。
        - 仅统计 department 非空的账号；结果按部门名升序。
      responses:
        '200':
          description: 成功，data 为部门名字符串数组
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: ok }
                  data:
                    type: array
                    items: { type: string }
                    example: [工程部, 市场部, 财务部]
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非 admin 角色（Admin access required）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /accounts/{id}:
    get:
      tags: [账号]
      summary: 账号详情
      description: |-
        【用途】管理员查看指定账号详情。

        【调用时机】管理后台点击某个成员查看/编辑前拉取详情时。

        【业务规则与注意事项】
        - 仅 admin 角色可调用，普通用户返回 403。
        - 返回 SafeUser 白名单字段，不含密码哈希。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 用户 ID（UUID）
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
      responses:
        '200':
          description: 成功，data 为 SafeUser
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: ok }
                  data:
                    $ref: '#/components/schemas/SafeUser'
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非 admin 角色（Admin access required）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: 账号不存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    delete:
      tags: [账号]
      summary: 软删除账号（仅 admin）
      description: |-
        【用途】管理员软删除（注销）一个账号：账号打 deleted_at 标记，物理数据与历史消息全部保留。

        【调用时机】管理后台「成员管理」对离职/异常账号执行注销时。

        【业务规则与注意事项】
        - 仅 admin 角色可调用；不能删除自己，也不能删除其他 admin 账号（均返回 400）。
        - 删除立即生效：该账号全部已签发 token 即刻失效（JWT 守卫实时查库校验），全端强制下线。
        - 历史消息保留，前端对 deleted_at 非空的发送者统一展示「已注销用户」。
        - 删除后同手机号可重新开通账号（新账号为全新身份，不继承旧数据）。
        - 账号列表（GET /accounts）默认不含已软删账号，传 show_deleted=1 可查看；审计日志留痕（action=delete_account）。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 账号 ID（UUID）
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
      responses:
        '200':
          description: 删除成功（message 为「账号已删除」），data 为 null
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 账号已删除 }
                  data: { nullable: true, example: null }
        '400':
          description: 不能删除自己 / 不能删除管理员账号 / 账号已被删除
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非 admin 角色（Admin access required）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: 账号不存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /accounts/{id}/reset-password:
    post:
      tags: [账号]
      summary: 重置指定账号密码
      description: |-
        【用途】管理员重置指定账号的密码。

        【调用时机】员工忘记密码找管理员重置；v1 旧数据迁移后也需通过此接口统一重置（旧版 bcrypt 与现版 Argon2 哈希不兼容）。

        【业务规则与注意事项】
        - 仅 admin 角色可调用，普通用户返回 403。
        - new_password 留空时自动生成 12 位随机密码并在响应中返回一次；指定时响应返回的即为所指定值。
        - 重置后该账号 force_change_pwd=true，员工下次登录将被强制改密。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 目标用户 ID（UUID）
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                new_password:
                  type: string
                  description: 指定新密码（可选）；留空则自动生成随机密码并在响应中返回一次
                  example: 'Reset@2026'
      responses:
        '200':
          description: 重置成功（message 为「密码重置成功」）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 密码重置成功 }
                  data:
                    type: object
                    properties:
                      new_password:
                        type: string
                        description: 新密码（未指定时为服务端生成的随机密码，仅此一次返回）
                        example: 'kX9#mQ2wE8rT'
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非 admin 角色（Admin access required）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: 账号不存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /accounts/{id}/toggle-status:
    post:
      tags: [账号]
      summary: 启用/停用账号；停用时删除该账号全部设备记录（全端强制下线）
      description: |-
        【用途】管理员启用或停用指定账号。

        【调用时机】员工离职（停用）或复职/误停恢复（启用）时。

        【业务规则与注意事项】
        - 仅 admin 角色可调用，普通用户返回 403。
        - 停用（status=disabled）时会删除该账号的全部设备记录；由于 JWT 守卫每个请求都实时校验账号状态与设备存在性，该账号所有已签发 token 立即失效，实现全端即时强制下线。
        - 停用账号再登录将返回 401「账号已被停用」。
        - 启用（status=active）仅恢复账号状态，不恢复已删除的设备记录，用户需重新登录。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 目标用户 ID（UUID）
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [active, disabled]
                  description: 目标状态：active 启用 / disabled 停用（停用即全端强制下线）
                  example: disabled
      responses:
        '200':
          description: 更新成功（message 为「状态更新成功」，data 为 null）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 状态更新成功 }
                  data: { type: object, nullable: true, example: null }
        '400':
          description: status 缺失或取值非法（仅允许 active/disabled）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非 admin 角色（Admin access required）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: 账号不存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |-
        JWT access token。登录（POST /auth/login）或刷新（POST /auth/refresh-token）后获得，放入请求头：
        Authorization: Bearer <access_token>
        默认有效期 7200 秒；每个受保护请求都会实时校验账号状态与设备存在性，账号停用或设备下线后 token 立即失效。
  schemas:
    Error:
      type: object
      description: 统一错误响应（code 为 HTTP 状态码；参数校验失败时 message 为多条错误以 '; ' 拼接）
      properties:
        code:
          type: integer
          description: HTTP 状态码（400/401/403/404/409/413/429/500）
          example: 400
        message:
          type: string
          description: 错误描述
          example: 手机号格式不正确
      required: [code, message]
    SafeUser:
      type: object
      description: 对外用户信息（字段白名单，任何接口不会返回 password_hash 等敏感字段）
      properties:
        id:
          type: string
          format: uuid
          description: 用户 ID（UUID v4）
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        phone:
          type: string
          description: 手机号（即用户名），1 开头 11 位数字
          example: '13800138000'
        display_name:
          type: string
          description: 显示名/姓名
          example: 张三
        avatar_url:
          type: string
          nullable: true
          description: 头像 URL，未设置为 null
          example: /uploads/9b1d2c3e-4f5a-4b6c-8d7e-9f0a1b2c3d4e.jpg
        signature:
          type: string
          nullable: true
          description: 个性签名，未设置为 null
          example: 保持热爱，奔赴山海
        department:
          type: string
          nullable: true
          description: 部门，未设置为 null
          example: 工程部
        role:
          type: string
          enum: [user, admin]
          description: 角色：user 普通用户 / admin 管理员（可调用账号管理接口）
          example: user
        status:
          type: string
          enum: [active, disabled]
          description: 账号状态：active 正常 / disabled 已停用（停用后 token 立即失效且无法登录）
          example: active
        force_change_pwd:
          type: boolean
          description: 是否需强制改密；true 时前端应强制跳转修改密码（新开通或被重置密码的账号为 true）
          example: false
        created_at:
          type: string
          format: date-time
          description: 创建时间（服务端 datetime）
          example: '2026-08-01T02:00:00.000Z'
        updated_at:
          type: string
          format: date-time
          description: 更新时间（服务端 datetime）
          example: '2026-08-20T09:30:00.000Z'
...
    BatchResult:
      type: object
      description: 批量开通账号的逐条结果
      properties:
        phone:
          type: string
          description: 手机号
          example: '13900139002'
        display_name:
          type: string
          description: 姓名
          example: 李四
        password:
          type: string
          description: 初始密码（成功时为服务端生成的随机密码，仅此一次返回；失败时为空字符串）
          example: 'kX9#mQ2wE8rT'
        success:
          type: boolean
          description: 该条是否开通成功
          example: true
        error:
          type: string
          description: 失败原因（仅失败项存在，如「该手机号已开通账号」）
          example: 该手机号已开通账号